add TimestampPattern

+++ src/Helper.php
@@ -19,30 +19,29 @@ class Helper
      * @param string $src
      * @param string $dst
      */
-    static function copyDirectory(
-        string $src,
-        string $dst
-    ): void {
+    static function copyDirectory(string $src, string $dst): void
+    {
+        if (is_link($src) || is_link($dst)) {
+            return;
+        }
+
         $dir = opendir($src);
-        mkdir($dst, 0777, true);
+        mkdir($dst, recursive: true);
         while (($file = readdir($dir)) !== false) {
-            if (
-                $file === static::CURRENT_DIRECTORY
-                || $file === static::PARENT_DIRECTORY
-            ) {
+            if ($file === static::CURRENT_DIRECTORY || $file === static::PARENT_DIRECTORY) {
                 continue;
             }
 
+            $newSrc = $src . DIRECTORY_SEPARATOR . $file;
+            $newDst = $dst . DIRECTORY_SEPARATOR . $file;
+
             if (is_dir($src . DIRECTORY_SEPARATOR . $file)) {
-                static::copyDirectory(
-                    $src . DIRECTORY_SEPARATOR . $file,
-                    $dst . DIRECTORY_SEPARATOR . $file
-                );
+                static::copyDirectory($newSrc, $newDst);
             } else {
-                copy(
-                    $src . DIRECTORY_SEPARATOR . $file,
-                    $dst . DIRECTORY_SEPARATOR . $file
-                );
+                if (is_link($newSrc) || is_link($newDst)) {
+                    continue;
+                }
+                copy($newSrc, $newDst);
             }
         }
         closedir($dir);
@@ -53,10 +52,14 @@ class Helper
      *
      * @param string $path
      */
-    static function removeDirectory(string $path)
+    static function removeDirectory(string $path): void
     {
         if (substr($path, -1) !== DIRECTORY_SEPARATOR) {
             $path .= DIRECTORY_SEPARATOR;
+        }
+
+        if (!is_dir($path) || is_link($path)) {
+            return;
         }
 
         $directory = dir($path);
@@ -66,7 +69,7 @@ class Helper
             }
 
             $filePath = $path . $file;
-            if (is_dir($filePath)) {
+            if (is_dir($filePath) && !is_link($filePath)) {
                 static::removeDirectory($filePath);
             } else {
                 unlink($filePath);
